+++ client/src/lib/parse/data/SetParser.ts
@@ -73,7 +73,7 @@ export default class SetParser extends EntityParser {
     const halLinks = this.getHalLinks()
     if (
       halLinks !== null &&
-      (halLinks.hasOwnProperty(`lux:setIncludedWorks`) ||
+      (halLinks.hasOwnProperty(`lux:setIncludedSets`) ||
         halLinks.hasOwnProperty(`lux:setIncludedItems`))
     ) {
       return true
@@ -109,7 +109,6 @@ export default class SetParser extends EntityParser {
     null | string | Array<any> | IContentWithLanguage
   > | null {
     const names = this.getNames()
-    const entityClass = this.getEntityClass('work')
     const itemType = this.getTypes()
     const identifiers = this.getIdentifiers()
     const setCreationEvent = this.getCreationEvent()
@@ -121,7 +120,6 @@ export default class SetParser extends EntityParser {
 
     const data: Record<string, any> = {
       names,
-      entityClass,
       itemType,
       identifiers,
       setCreationEvent,